+++ frontend-nodejs/src/app/(auth)/register/components/BookmarkError.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+export const BookmarkError = ({ message }: { message: string }) => {
+  return (
+    <div className="relative mt-1 mx-auto animate-bounce">
+      <div className="absolute w-0 h-0 border-l-[10px] border-l-transparent border-r-[10px] border-r-transparent border-t-[15px] border-t-red-500 left-4 -top-1 z-10" />
+      <div className="bg-red-500 text-white px-4 py-1 rounded text-sm font-bold shadow-md">
+        {message}
+      </div>
+    </div>
+  );
+};
